Handling zero PDF items
I have added `KeepIndexesForZeroPDF` and `IgnoreIndexesForZeroPDF` defining how PickVector will handle pdf
values that are 0 (for example, someone turns off a region). See README.md for examples.

This use-case was not implemented before, so I had to make the necessary modifications in the code.
I am releasing a new minor version as well.

Signed-off-by: kuritka <kuritka@gmail.com>